paddyai-backend/requirements.txt
flask
flask-cors
scikit-learn
numpy
pillow
pandas
gunicorn
gdown
